pet-adoption-backend/requirements.txt
# Core Flask
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-CORS==5.0.0
Flask-JWT-Extended==4.6.0

# Database
SQLAlchemy==2.0.23
alembic==1.13.0

# Security & Authentication
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0

# Image Handling
Pillow==10.1.0

# Server
gunicorn==21.2.0
werkzeug==3.0.1

# Utilities
email-validator==2.1.0.post1
python-dateutil==2.8.2

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0